feat: добавлены аутентификация и авторизация, разделение на роли и права (#46)

+++ app/routers/projects/projects.py
@@ -1,23 +1,24 @@
 from typing import Annotated, Optional, Sequence
 from uuid import UUID
 
-from fastapi import APIRouter, Query
+from fastapi import APIRouter, Depends, HTTPException, Query, status
 
+from app.core.config import settings
+from app.dependencies.auth import CurrentStudentDep, MemberRole, ProjectRoleDep
 from app.dependencies.services import (
-    ProjectServiceDep,
     ProjectMemberServiceDep,
+    ProjectServiceDep,
 )
-from app.models.project import (
+from app.models.projects.project import (
     ProjectCreate,
-    ProjectMemberCreate,
-    ProjectMemberPublic,
     ProjectPublic,
     ProjectUpdate,
 )
-from app.routers.projects.projects import (
-    projects_teams,
+from app.models.projects.project_member import (
+    ProjectMemberCreate,
+    ProjectMemberPublic,
 )
-from app.routers.projects.projects import projects_members, projects_team
+from app.routers.projects import projects_members, projects_team, projects_teams
 from app.schemas.projects import ProjectFilters
 
 router = APIRouter(
@@ -30,7 +31,7 @@ router.include_router(projects_teams.router)
 router.include_router(projects_members.router)
 
 
-@router.get('/')
+@router.get('/', dependencies=[Depends(CurrentStudentDep)])
 async def get_projects(
     project_service: ProjectServiceDep,
     filters: Annotated[ProjectFilters, Query()]
@@ -38,15 +39,14 @@ async def get_projects(
     return await project_service.get_projects(filters)
 
 
-@router.post('/')
+@router.post('/', dependencies=[Depends(CurrentStudentDep)])
 async def create_project(
-    project_service: ProjectServiceDep,
-    project_create: ProjectCreate
+    project_service: ProjectServiceDep, project_create: ProjectCreate
 ) -> ProjectPublic:
     return await project_service.create_project(project_create)
 
 
-@router.post('/join')
+@router.post('/join', dependencies=[Depends(CurrentStudentDep)])
 async def join_project(
     project_member_service: ProjectMemberServiceDep,
     pm_create: ProjectMemberCreate
@@ -54,26 +54,41 @@ async def join_project(
     return await project_member_service.add_member_to_project(pm_create)
 
 
-@router.get('/{project_id}')
+@router.get('/{project_id}', dependencies=[Depends(CurrentStudentDep)])
 async def get_project(
-    project_service: ProjectServiceDep,
-    project_id: UUID
+    project_service: ProjectServiceDep, project_id: UUID
 ) -> Optional[ProjectPublic]:
     return await project_service.get_project(project_id)
 
 
 @router.put('/{project_id}')
 async def update_project(
+    student: CurrentStudentDep,
+    project_role: ProjectRoleDep,
     project_service: ProjectServiceDep,
     project_update: ProjectUpdate,
-    project_id: UUID
+    project_id: UUID,
 ) -> Optional[ProjectPublic]:
+    if (
+        student.role == settings.role.default_user_role_code
+        and project_role != MemberRole.TEAMLEAD
+    ):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
     return await project_service.update_project(project_update, project_id)
 
 
 @router.delete('/{project_id}')
 async def detele_project(
+    student: CurrentStudentDep,
+    project_role: ProjectRoleDep,
     project_member_service: ProjectServiceDep,
-    project_id: UUID
+    project_id: UUID,
 ) -> Optional[ProjectPublic]:
+    if (
+        student.role == settings.role.default_user_role_code
+        and project_role != MemberRole.TEAMLEAD
+    ):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
     return await project_member_service.delete_project(project_id)
