+++ app/routers/projects/projects_team.py
@@ -1,18 +1,22 @@
 from typing import Optional, Sequence
 from uuid import UUID
 
-from fastapi import APIRouter
+from fastapi import APIRouter, Depends, HTTPException, status
 
+from app.core.config import settings
+from app.dependencies.auth import CurrentStudentDep, MemberRole, ProjectRoleDep
 from app.dependencies.services import (
     TeamMembershipServiceDep,
     TeamServiceDep,
 )
-from app.models.projects.project import (
+from app.models.projects.team import (
+    TeamPublic,
+    TeamUpdate,
+)
+from app.models.projects.team_membership import (
     TeamMembershipCreate,
     TeamMembershipPublic,
     TeamMembershipUpdate,
-    TeamPublic,
-    TeamUpdate,
 )
 from app.schemas.projects import TeamMembershipFilters
 
@@ -24,61 +28,101 @@ router = APIRouter(
 
 @router.get('/')
 async def get_team(
+    student: CurrentStudentDep,
+    project_role: ProjectRoleDep,
     team_service: TeamServiceDep,
-    project_id: UUID,
     team_id: UUID
 ) -> Optional[TeamPublic]:
+    if (
+        student.role == settings.role.default_user_role_code
+        and project_role == MemberRole.OTHER
+    ):
+        raise HTTPException(status=status.HTTP_403_FORBIDDEN)
+
     return await team_service.get_team(team_id)
 
 
 @router.put('/')
 async def update_team(
+    student: CurrentStudentDep,
+    project_role: ProjectRoleDep,
     team_service: TeamServiceDep,
-    project_id: UUID,
     team_id: UUID,
-    team_update: TeamUpdate
+    team_update: TeamUpdate,
 ) -> Optional[TeamPublic]:
+    if (
+        student.role == settings.role.default_user_role_code
+        and project_role != MemberRole.TEAMLEAD
+    ):
+        raise HTTPException(status=status.HTTP_403_FORBIDDEN)
+
     return await team_service.update_team(team_id, team_update)
 
 
 @router.delete('/')
 async def delete_team(
+    student: CurrentStudentDep,
+    project_role: ProjectRoleDep,
     team_service: TeamServiceDep,
-    project_id: UUID,
-    team_id: UUID
+    team_id: UUID,
 ) -> Optional[TeamPublic]:
+    if (
+        student.role == settings.role.default_user_role_code
+        and project_role != MemberRole.TEAMLEAD
+    ):
+        raise HTTPException(status=status.HTTP_403_FORBIDDEN)
+
     return await team_service.delete_team(team_id)
 
 
 @router.get('/members')
 async def get_team_members(
+    student: CurrentStudentDep,
+    project_role: ProjectRoleDep,
     team_membership_service: TeamMembershipServiceDep,
-    project_id: UUID,
     team_id: UUID,
-    filters: TeamMembershipFilters
+    filters: TeamMembershipFilters,
 ) -> Sequence[TeamMembershipPublic]:
+    if (
+        student.role == settings.role.default_user_role_code
+        and project_role == MemberRole.OTHER
+    ):
+        raise HTTPException(status=status.HTTP_403_FORBIDDEN)
     filters.team_id = team_id
     return await team_membership_service.get_members(filters)
 
 
 @router.post('/members')
 async def create_team_member(
+    student: CurrentStudentDep,
+    project_role: ProjectRoleDep,
     team_membership_service: TeamMembershipServiceDep,
-    project_id: UUID,
     team_id: UUID,
-    tm_create: TeamMembershipCreate
+    tm_create: TeamMembershipCreate,
 ) -> TeamMembershipPublic:
+    if (
+        student.role == settings.role.default_user_role_code
+        and project_role != MemberRole.TEAMLEAD
+    ):
+        raise HTTPException(status=status.HTTP_403_FORBIDDEN)
+
     tm_create.team_id = team_id
     return await team_membership_service.create_membership(tm_create)
 
 
 @router.get('/members/{student_id}')
 async def get_member(
+    student: CurrentStudentDep,
+    project_role: ProjectRoleDep,
     team_membership_service: TeamMembershipServiceDep,
-    project_id: UUID,
     team_id: UUID,
-    student_id: UUID
+    student_id: UUID,
 ) -> Optional[TeamMembershipPublic]:
+    if (
+        student.role == settings.role.default_user_role_code
+        and project_role == MemberRole.OTHER
+    ):
+        raise HTTPException(status=status.HTTP_403_FORBIDDEN)
     filters = TeamMembershipFilters
     filters.team_id = team_id
     filters.project_member_id = student_id
@@ -87,26 +131,42 @@ async def get_member(
 
 @router.put('/members/{student_id}')
 async def update_team_member(
+    student: CurrentStudentDep,
+    project_role: ProjectRoleDep,
     team_membership_service: TeamMembershipServiceDep,
-    project_id: UUID,
     team_id: UUID,
     student_id: UUID,
-    tm_update: TeamMembershipUpdate
+    tm_update: TeamMembershipUpdate,
 ) -> Optional[TeamMembershipPublic]:
+    if (
+        student.role == settings.role.default_user_role_code
+        and project_role == MemberRole.OTHER
+    ):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
+    if student.id != student_id or student.role != MemberRole.TEAMLEAD:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
     filters = TeamMembershipFilters()
     filters.team_id = team_id
     filters.project_member_id = student_id
-    return await team_membership_service.update_membership(
-        filters, tm_update)
+    return await team_membership_service.update_membership(filters, tm_update)
 
 
 @router.delete('/members/{student_id}')
 async def delete_membership(
+    student: CurrentStudentDep,
+    project_role: ProjectRoleDep,
     team_membership_service: TeamMembershipServiceDep,
-    project_id: UUID,
     team_id: UUID,
-    student_id: UUID
+    student_id: UUID,
 ) -> Optional[TeamMembershipPublic]:
+    if (
+        student.role == settings.role.default_user_role_code
+        and project_role != MemberRole.TEAMLEAD
+    ):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
     filters = TeamMembershipFilters()
     filters.team_id = team_id
     filters.project_member_id = student_id
